refactor CachedNauticalEarthFile back into Natural EArth plugin
