+++ org.mwc.cmap.NaturalEarth/src/org/mwc/cmap/naturalearth/Activator.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.jface.resource.ImageDescriptor;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
-import org.mwc.cmap.gt2plot.data.CachedNauticalEarthFile;
+import org.mwc.cmap.naturalearth.data.CachedNauticalEarthFile;
 import org.mwc.cmap.naturalearth.preferences.PreferenceConstants;
 import org.mwc.cmap.naturalearth.view.NEFeatureSet;
 import org.mwc.cmap.naturalearth.view.NEFeatureStyle;
@@ -144,6 +144,32 @@ public class Activator extends AbstractUIPlugin
 		}
 
 		return res;
+	}
+	
+	/**
+	 * error logging utility
+	 * 
+	 * @param severity
+	 *          the severity; one of <code>OK</code>, <code>ERROR</code>,
+	 *          <code>INFO</code>, <code>WARNING</code>, or <code>CANCEL</code>
+	 * @param message
+	 *          a human-readable message, localized to the current locale
+	 * @param exception
+	 *          a low-level exception, or <code>null</code> if not applicable
+	 */
+	public static void logError(final int severity, final String message, final Throwable exception)
+	{
+		final Activator singleton = getDefault();
+		if (singleton != null)
+		{
+			final Status stat = new Status(severity, "org.mwc.cmap.NaturalEarth", Status.OK,
+					message, exception);
+			singleton.getLog().log(stat);
+		}
+
+		// also throw it to the console
+		if (exception != null)
+			exception.printStackTrace();
 	}
 	
 	public static NEResolution getStyleFor(double curScale)
