+++ org.mwc.cmap.geotools/META-INF/MANIFEST.MF
@@ -23,3 +23,191 @@ Bundle-ClassPath: libs/commons-pool-1.5.4.jar,
  libs/jts-1.13.jar,
  libs/vecmath-1.3.2.jar,
  .
+Export-Package: com.vividsolutions.jts,
+ com.vividsolutions.jts.algorithm,
+ com.vividsolutions.jts.algorithm.distance,
+ com.vividsolutions.jts.algorithm.locate,
+ com.vividsolutions.jts.algorithm.match,
+ com.vividsolutions.jts.awt,
+ com.vividsolutions.jts.densify,
+ com.vividsolutions.jts.geom,
+ com.vividsolutions.jts.geom.impl,
+ com.vividsolutions.jts.geom.prep,
+ com.vividsolutions.jts.geom.util,
+ com.vividsolutions.jts.geomgraph,
+ com.vividsolutions.jts.geomgraph.index,
+ com.vividsolutions.jts.index,
+ com.vividsolutions.jts.index.bintree,
+ com.vividsolutions.jts.index.chain,
+ com.vividsolutions.jts.index.intervalrtree,
+ com.vividsolutions.jts.index.kdtree,
+ com.vividsolutions.jts.index.quadtree,
+ com.vividsolutions.jts.index.strtree,
+ com.vividsolutions.jts.index.sweepline,
+ com.vividsolutions.jts.io,
+ com.vividsolutions.jts.io.gml2,
+ com.vividsolutions.jts.linearref,
+ com.vividsolutions.jts.math,
+ com.vividsolutions.jts.noding,
+ com.vividsolutions.jts.noding.snapround,
+ com.vividsolutions.jts.operation,
+ com.vividsolutions.jts.operation.buffer,
+ com.vividsolutions.jts.operation.buffer.validate,
+ com.vividsolutions.jts.operation.distance,
+ com.vividsolutions.jts.operation.distance3d,
+ com.vividsolutions.jts.operation.linemerge,
+ com.vividsolutions.jts.operation.overlay,
+ com.vividsolutions.jts.operation.overlay.snap,
+ com.vividsolutions.jts.operation.overlay.validate,
+ com.vividsolutions.jts.operation.polygonize,
+ com.vividsolutions.jts.operation.predicate,
+ com.vividsolutions.jts.operation.relate,
+ com.vividsolutions.jts.operation.union,
+ com.vividsolutions.jts.operation.valid,
+ com.vividsolutions.jts.planargraph,
+ com.vividsolutions.jts.planargraph.algorithm,
+ com.vividsolutions.jts.precision,
+ com.vividsolutions.jts.shape,
+ com.vividsolutions.jts.shape.fractal,
+ com.vividsolutions.jts.shape.random,
+ com.vividsolutions.jts.simplify,
+ com.vividsolutions.jts.triangulate,
+ com.vividsolutions.jts.triangulate.quadedge,
+ com.vividsolutions.jts.util,
+ com.vividsolutions.jtsexample.geom,
+ com.vividsolutions.jtsexample.geom.prep,
+ com.vividsolutions.jtsexample.io.gml2,
+ com.vividsolutions.jtsexample.linearref,
+ com.vividsolutions.jtsexample.operation.distance,
+ com.vividsolutions.jtsexample.operation.linemerge,
+ com.vividsolutions.jtsexample.operation.polygonize,
+ com.vividsolutions.jtsexample.precision,
+ com.vividsolutions.jtsexample.technique,
+ org.geotools.console,
+ org.geotools.data,
+ org.geotools.data.collection,
+ org.geotools.data.crs,
+ org.geotools.data.directory,
+ org.geotools.data.memory,
+ org.geotools.data.ows,
+ org.geotools.data.shapefile,
+ org.geotools.data.shapefile.dbf,
+ org.geotools.data.shapefile.fid,
+ org.geotools.data.shapefile.files,
+ org.geotools.data.shapefile.index,
+ org.geotools.data.shapefile.index.quadtree,
+ org.geotools.data.shapefile.index.quadtree.fs,
+ org.geotools.data.shapefile.shp,
+ org.geotools.data.shapefile.shp.xml,
+ org.geotools.data.simple,
+ org.geotools.data.sort,
+ org.geotools.data.store,
+ org.geotools.data.view,
+ org.geotools.factory,
+ org.geotools.feature,
+ org.geotools.feature.collection,
+ org.geotools.feature.simple,
+ org.geotools.feature.type,
+ org.geotools.feature.visitor,
+ org.geotools.filter,
+ org.geotools.filter.capability,
+ org.geotools.filter.expression,
+ org.geotools.filter.function,
+ org.geotools.filter.function.math,
+ org.geotools.filter.function.string,
+ org.geotools.filter.identity,
+ org.geotools.filter.spatial,
+ org.geotools.filter.temporal,
+ org.geotools.filter.visitor,
+ org.geotools.geometry,
+ org.geotools.geometry.coordinatesequence,
+ org.geotools.geometry.jts,
+ org.geotools.geometry.jts.coordinatesequence,
+ org.geotools.geometry.text,
+ org.geotools.gml,
+ org.geotools.gml.producer,
+ org.geotools.io,
+ org.geotools.math,
+ org.geotools.measure,
+ org.geotools.metadata,
+ org.geotools.metadata.iso,
+ org.geotools.metadata.iso.citation,
+ org.geotools.metadata.iso.constraint,
+ org.geotools.metadata.iso.content,
+ org.geotools.metadata.iso.distribution,
+ org.geotools.metadata.iso.extent,
+ org.geotools.metadata.iso.identification,
+ org.geotools.metadata.iso.lineage,
+ org.geotools.metadata.iso.maintenance,
+ org.geotools.metadata.iso.quality,
+ org.geotools.metadata.iso.spatial,
+ org.geotools.metadata.sql,
+ org.geotools.nature,
+ org.geotools.ows,
+ org.geotools.parameter,
+ org.geotools.referencing,
+ org.geotools.referencing.crs,
+ org.geotools.referencing.cs,
+ org.geotools.referencing.datum,
+ org.geotools.referencing.factory,
+ org.geotools.referencing.factory.epsg,
+ org.geotools.referencing.factory.gridshift,
+ org.geotools.referencing.factory.wms,
+ org.geotools.referencing.operation,
+ org.geotools.referencing.operation.builder,
+ org.geotools.referencing.operation.matrix,
+ org.geotools.referencing.operation.projection,
+ org.geotools.referencing.operation.transform,
+ org.geotools.referencing.wkt,
+ org.geotools.renderer,
+ org.geotools.resources,
+ org.geotools.resources.geometry,
+ org.geotools.resources.i18n,
+ org.geotools.styling,
+ org.geotools.styling.visitor,
+ org.geotools.temporal,
+ org.geotools.temporal.object,
+ org.geotools.temporal.reference,
+ org.geotools.text,
+ org.geotools.util,
+ org.geotools.util.logging,
+ org.geotools.xml.transform,
+ org.opengis.annotation,
+ org.opengis.coverage,
+ org.opengis.coverage.grid,
+ org.opengis.coverage.processing,
+ org.opengis.feature,
+ org.opengis.feature.simple,
+ org.opengis.feature.type,
+ org.opengis.filter,
+ org.opengis.filter.capability,
+ org.opengis.filter.expression,
+ org.opengis.filter.identity,
+ org.opengis.filter.sort,
+ org.opengis.filter.spatial,
+ org.opengis.filter.temporal,
+ org.opengis.geometry,
+ org.opengis.geometry.aggregate,
+ org.opengis.geometry.complex,
+ org.opengis.geometry.coordinate,
+ org.opengis.geometry.primitive,
+ org.opengis.metadata,
+ org.opengis.metadata.citation,
+ org.opengis.metadata.constraint,
+ org.opengis.metadata.content,
+ org.opengis.metadata.distribution,
+ org.opengis.metadata.extent,
+ org.opengis.metadata.identification,
+ org.opengis.metadata.lineage,
+ org.opengis.metadata.maintenance,
+ org.opengis.metadata.quality,
+ org.opengis.metadata.spatial,
+ org.opengis.parameter,
+ org.opengis.referencing,
+ org.opengis.referencing.crs,
+ org.opengis.referencing.cs,
+ org.opengis.referencing.datum,
+ org.opengis.referencing.operation,
+ org.opengis.style,
+ org.opengis.temporal,
+ org.opengis.util
